+++ todo-app-frontend/src/components/todo/TodoApp.jsx
@@ -1,67 +1,33 @@
-import { useState } from "react";
+import LoginComponent from "./LoginComponent";
+import WelcomeComponent from "./WelcomeComponent";
+import {BrowserRouter as Router, Routes, Route} from 'react-router-dom';
+import withNavigation from "./withNavigation";
+import withParams from "./withParams";
 
 const TodoApp = () => {
-    return (
-        <div className="TodoApp">
-            <LoginComponent />
-        </div>
-    )
-}
-
-export default TodoApp;
-
-const LoginComponent = () => {
-    const [username, setUsername] = useState("");
-    const [password, setPassword] = useState("");
-    const [hasLoginFailed, setHasLoginFailed] = useState(false);
-    const [showLoginSuccess, setShowLoginSuccess] = useState(false);
-
-    const handleChange = (event) => {
-        switch(event.target.name) {
-            case "username": 
-                setUsername(event.target.value);
-                break;
-            case "password": 
-                setPassword(event.target.value);
-                break;
-            default: 
-                break;
-        }  
-    }
-
-    const handleLogin = () => {
-        // jon, dummy
-        if (username === "jon" && password === "dummy") {
-            console.log("Successful login");
-            setShowLoginSuccess(true);
-            setHasLoginFailed(false);
-        }
-        else {
-            console.log("Invalid Credentials at login");
-            setShowLoginSuccess(false);
-            setHasLoginFailed(true);
-        }
-    }
+    const LoginComponentWithNavigation = withNavigation(LoginComponent);
+    const WelcomeComponentWithParams = withParams(WelcomeComponent);
 
     return (
-      <div>
-        {hasLoginFailed && <div>Invalid Credentials</div>}
-        {showLoginSuccess && <div>Successful Login</div>}
-        Username:
-        <input
-          type="text"
-          name="username"
-          value={username}
-          onChange={handleChange}
-        />
-        Password:
-        <input
-          type="password"
-          name="password"
-          value={password}
-          onChange={handleChange}
-        />
-        <button onClick={handleLogin}>Login</button>
+      <div className="TodoApp">
+        <Router>
+          <Routes>
+            <Route path="/" exact element={<LoginComponentWithNavigation />} />
+            <Route path="/login" element={<LoginComponentWithNavigation />} />
+            <Route path="/welcome/:name" element={<WelcomeComponentWithParams />} />
+            <Route
+              path="*"
+              element={
+                <main style={{ padding: "1rem" }}>
+                  <p>An Error Occurred.</p>
+                  <p>There's nothing here!</p>
+                </main>
+              }
+            />
+          </Routes>
+        </Router>
       </div>
     );
 }
+
+export default TodoApp;
