+++ todo-app-frontend/src/components/todo/WelcomeComponent.jsx
@@ -1,0 +1,7 @@
+
+
+const WelcomeComponent = ({ params: { name } }) => {
+    return <div>Welcome {name} !</div>
+}
+
+export default WelcomeComponent;
