+++ todo-app-frontend/src/components/todo/withNavigation.jsx
@@ -1,0 +1,7 @@
+import { useNavigate } from "react-router-dom";
+
+const withNavigation = (Component) => {
+  return (props) => <Component {...props} navigate={useNavigate()} />;
+};
+
+export default withNavigation;
