+++ todo-app-frontend/src/components/todo/withParams.jsx
@@ -1,0 +1,7 @@
+import { useParams } from "react-router-dom";
+
+const withParams = (Component) => {
+    return (props) => <Component {...props} params={useParams()} />;
+}
+
+export default withParams;
